Add __call() to XcacheClassLoader
XcacheClassLoader does not implement __call() to forward methods (such as addPsr4()) to the decorated object. I think this is an oversight, although I might be missing something. I just copied the code over from ApcClassLoader, which does implement __call().

This allows to use XcacheClassLoader with Drupal 8, as the addPsr4() method is used there.